REF: Rename recursive least squares module.
