+++ statsmodels/regression/recursive_ls.py
